chore: internal restructuring to improve performance
