+++ test/src/commonMain/kotlin/pro/respawn/flowmvi/test/plugin/TestPipelineContext.kt
@@ -6,7 +6,7 @@ import kotlinx.atomicfu.atomic
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
 import pro.respawn.flowmvi.api.DelicateStoreApi
-import pro.respawn.flowmvi.api.ExperimentalStoreApi
+import pro.respawn.flowmvi.annotation.ExperimentalStoreApi
 import pro.respawn.flowmvi.api.MVIAction
 import pro.respawn.flowmvi.api.MVIIntent
 import pro.respawn.flowmvi.api.MVIState
